+++ src/redturtle/prenotazioni/tests/test_delete_booking.py
@@ -85,7 +85,7 @@ class TestDeleteBooking(unittest.TestCase):
         self.assertRaises(NotFound, self.view.do_delete)
 
     def test_admin_can_delete_booking(self):
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today + timedelta(1),  # tomorrow
                 "booking_type": "Type A",
@@ -103,7 +103,7 @@ class TestDeleteBooking(unittest.TestCase):
 
     @unittest.skip("wip")
     def test_anon_cant_delete_other_user_booking(self):
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today + timedelta(1),  # tomorrow
                 "booking_type": "Type A",
@@ -120,7 +120,7 @@ class TestDeleteBooking(unittest.TestCase):
 
     def test_anon_can_delete_anon_booking(self):
         logout()
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today + timedelta(1),  # tomorrow
                 "booking_type": "Type A",
@@ -144,7 +144,7 @@ class TestDeleteBooking(unittest.TestCase):
 
     def test_user_can_delete_his_booking(self):
         login(self.portal, "user")
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today + timedelta(1),  # tomorrow
                 "booking_type": "Type A",
@@ -169,7 +169,7 @@ class TestDeleteBooking(unittest.TestCase):
     @unittest.skip("wip")
     def test_user_cant_delete_other_user_booking(self):
         login(self.portal, "user")
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today + timedelta(1),  # tomorrow
                 "booking_type": "Type A",
@@ -194,7 +194,7 @@ class TestDeleteBooking(unittest.TestCase):
     @unittest.skip("wip")
     def test_user_cant_delete_anon_booking(self):
         logout()
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today + timedelta(1),  # tomorrow
                 "booking_type": "Type A",
@@ -219,7 +219,7 @@ class TestDeleteBooking(unittest.TestCase):
         )
 
     def test_cant_delete_past_booking(self):
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today - timedelta(1),  # yesterday
                 "booking_type": "Type A",
@@ -245,7 +245,7 @@ class TestDeleteBooking(unittest.TestCase):
         )
 
     def test_cant_delete_today_booking(self):
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today,
                 "booking_type": "Type A",