+++ CHANGES.rst
@@ -19,6 +19,9 @@ Changelog
 - Booking Managers bypass futureDays checks when create new bookings.
   [cekk]
 
+- Remove duplicated method in booker: create. Now there is only `book`.
+  [cekk]
+
 2.2.5 (2023-12-04)
 ------------------